REF: Refactor Date/TimeLikeOps
Moves the implementation / inheritence from indexes to arrays.
